@@ -8,7 +8,7 @@ dynamic = ["version"]
 description = "A Sphinx extension for producing exercises and solutions."
 readme = "README.md"
 license = { file = "LICENSE" }
-requires-python = ">=3.9"
+requires-python = ">=3.11"
 authors = [
     { name = "QuantEcon", email = "admin@quantecon.org" },
 ]
@@ -22,10 +22,9 @@ classifiers = [
     "Natural Language :: English",
     "Operating System :: OS Independent",
     "Programming Language :: Python",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Documentation",
     "Topic :: Documentation :: Sphinx",
     "Topic :: Software Development :: Documentation",
@@ -34,7 +33,7 @@ classifiers = [
 ]
 dependencies = [
     "sphinx-book-theme",
-    "sphinx>=5",
+    "sphinx>=6.1",
 ]
 
 [project.optional-dependencies]
@@ -49,20 +48,20 @@ code_style = [
     "pre-commit",
 ]
 rtd = [
-    "myst-nb~=1.0.0",
+    "myst-nb>=1.1.0",
     "sphinx-book-theme",
     "sphinx_togglebutton",
-    "sphinx>=5,<8",
+    "sphinx>=6.1,<9",
 ]
 testing = [
     "beautifulsoup4",
     "coverage",
-    "matplotlib==3.8.*",
-    "myst-nb~=1.0.0",
+    "matplotlib>=3.8",
+    "myst-nb>=1.1.0",
     "pytest-cov",
     "pytest-regressions",
-    "pytest~=8.0.0",
-    "sphinx>=5,<8",
+    "pytest>=8.0",
+    "sphinx>=6.1,<9",
     "texsoup",
     "defusedxml", # Required by sphinx-testing
 ]
