@@ -1,3 +1,5 @@
+from pathlib import Path
+
 import sphinx.addnodes as sphinx_nodes
 from sphinx.transforms.post_transforms import SphinxPostTransform
 from sphinx.util import logging
@@ -195,7 +197,7 @@ class ResolveTitlesInSolutions(SphinxPostTransform):
                     except AttributeError:
                         docname = self.env.docname  # for builder such as JupyterBuilder that don't support current_docname
                 docpath = self.env.doc2path(docname)
-                path = docpath[: docpath.rfind(".")]
+                path = str(Path(docpath).with_suffix(""))
                 msg = f"undefined label: {target_label}"
                 logger.warning(msg, location=path, color="red")
                 return
