@@ -1,9 +1,9 @@
 import json
-import os
 from pathlib import Path
 import subprocess
 
 MESSAGE_CATALOG_NAME = "exercise"
+
 
 def convert_json(folder=None):
     folder = folder or Path(__file__).parent
@@ -19,7 +19,13 @@ def convert_json(folder=None):
         english = data[0]["text"]
         for item in data[1:]:
             language = item["symbol"]
-            out_path = folder / "locales" / language / "LC_MESSAGES" / f"{MESSAGE_CATALOG_NAME}.po"
+            out_path = (
+                folder
+                / "locales"
+                / language
+                / "LC_MESSAGES"
+                / f"{MESSAGE_CATALOG_NAME}.po"
+            )
             if not out_path.parent.exists():
                 out_path.parent.mkdir(parents=True)
             if not out_path.exists():
@@ -47,9 +53,9 @@ msgstr ""
         subprocess.check_call(
             [
                 "msgfmt",
-                os.path.abspath(path),
+                str(path.resolve()),
                 "-o",
-                os.path.abspath(path.parent / f"{MESSAGE_CATALOG_NAME}.mo"),
+                str((path.parent / f"{MESSAGE_CATALOG_NAME}.mo").resolve()),
             ]
         )
 
