@@ -1,5 +1,8 @@
 from bs4 import BeautifulSoup
 import pytest
+import sphinx
+
+SPHINX_VERSION = f".sphinx{sphinx.version_info[0]}"
 
 
 @pytest.mark.sphinx("html", testroot="mybook")
@@ -29,7 +32,7 @@ def test_reference(app, idir, file_regression):
         excs += f"{ref}\n"
 
     file_regression.check(
-        str(excs[:-1]), basename=idir.split(".")[0], extension=".html"
+        str(excs[:-1]), basename=idir.split(".")[0], extension=f"{SPHINX_VERSION}.html"
     )
 
 
